Fixed type

+++ src/disp/display.cc
@@ -2196,9 +2196,9 @@ void Display::drawSolid(Solid const& obj)
         gym::ref_view();
         gym::enableLighting();
         gym::disableCullFace();
-        const size_t off = DIM+1; // index of point to be linked
-        const size_t sup = std::min(obj.lastPoint(), 2*off+1); // last point to be linked
-        for ( size_t inx = off; inx <= sup; inx += off+1 )
+        const unsigned off = DIM+1; // index of point to be linked
+        const unsigned sup = std::min(obj.lastPoint(), 2*off+1); // last point to be linked
+        for ( unsigned inx = off; inx <= sup; inx += off+1 )
         {
             gym::stretchAlignZ(obj.posPoint(inx), twi->posPoint(inx), rad);
             gym::color_both(lor, 1);
